% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cooksd.R
\name{cooksd}
\alias{cooksd}
\title{Evaluates Cook's Distance.}
\usage{
cooksd(fn, formula, data, family)
}
\arguments{
\item{fn}{The function to use when creating the model. Either "glm" or "gamlss".}

\item{formula}{The formula to use in the model.}

\item{data}{The dataframe.}

\item{family}{The distribution family. E.g. "poisson" if using glm. PO or NBI if using gamlss.}
}
\value{
A vector of distances
}
\description{
Evaluates Cook's Distance.
}
\note{
Although lm/glm stores the original data in the model at m$data, that is not the case in GAMLSS.
Therefore, unlike the function cooks.distance which only requires a model parameter, this function requires
the formula and data.
}
\examples{
cooksd(glm, f, df, "poisson")
cooksd(gamlss, f, df, PO)
 cooksd(gamlss, f, df, NBI)

}
\references{
https://www.ime.usp.br/~abe/lista/pdf1USQwcGBX1.pdf

Robust Diagnostic Regression Analysis, Anthony Atkinson and Marco Riani
}
